Lesson 2

+++ lesson1/triangle_area.rb
@@ -1,7 +1,7 @@
 puts "Введите длину основания треугольника:"
-base = gets.to_i
+base = gets.to_f
 puts "Введите высоту треугольника:"
-heigth = gets.to_i
+heigth = gets.to_f
 
 triangle_area = 0.5 * base * heigth
 puts "Площадь треугольника равна #{triangle_area}"
